Move archive fragment related classes to fragments.archive
