+++ babble/src/main/java/io/mosaicnetworks/babble/fragments/archive/ArchivedGroupsViewModelFactory.java
@@ -22,53 +22,31 @@
  * SOFTWARE.
  */
 
-package io.mosaicnetworks.babble.configure;
+package io.mosaicnetworks.babble.fragments.archive;
 
-import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModelProvider;
 
-import io.mosaicnetworks.babble.node.ConfigDirectory;
 import io.mosaicnetworks.babble.node.ConfigManager;
 
 /**
- * This class extends {@link ViewModel} and exposes methods to populate
- * the list of archive groups.
+ * Factory class to create {@link ArchivedGroupsViewModel} instance
  */
-public class ArchivedGroupsViewModel extends ViewModel {
-
-    private MutableLiveData<SelectableData<ConfigDirectory>> mArchivedList;
+public class ArchivedGroupsViewModelFactory implements ViewModelProvider.Factory {
     private ConfigManager mConfigManager;
 
     /**
-     * Constructor that set the instance of {@link ConfigManager} used to populate
-     * the archive group list.
-     *
-     * @param configManager a ConfigManager instance
+     * Constructor. Needs {@link ConfigManager} instance to pass to {@link ArchivedGroupsViewModel}
+     * @param configManager
      */
-    public ArchivedGroupsViewModel(ConfigManager configManager) {
-        super();
+    public ArchivedGroupsViewModelFactory(ConfigManager configManager) {
         mConfigManager = configManager;
-
-        mArchivedList = new MutableLiveData<>();
-        loadArchiveList();
     }
 
-    /**
-     * This method populates the mArchivedList global with Archive Group data
-     */
-    public void loadArchiveList() {
-        SelectableData<ConfigDirectory> data = new SelectableData<>();
-        data.addAll(mConfigManager.getDirectories());
-        mArchivedList.setValue(data);
-    }
 
-    /**
-     * Getter to retrieve the list populated by {@link #loadArchiveList()}
-     * @return
-     */
-    public MutableLiveData<SelectableData<ConfigDirectory>> getArchivedList() {
-        return mArchivedList;
+    @Override
+    @SuppressWarnings("unchecked")  //unchecked cast warning caused by the use of generics.
+    public <T extends ViewModel> T create(Class<T> modelClass) {
+        return (T) new ArchivedGroupsViewModel(mConfigManager);
     }
-
 }
-
